Handling HTTP request lifecycle events and error handling
The lifecycle structure of an HTTP request has already been defined, so all events are being emitted to the server based on the progress of the HTTP request.

Handling errors when establishing the connection with the server.

Listening to events emitted by the server during the HTTP request was implemented, however much of the logic to be implemented is unfinished for now.