Data transmission settings in HTTP request
Implemented aborting connections when determining that the websocket (tunnel) connection to the backend has been lost.

Implemented log every HTTP request completed (or rejected as the case may be).

Abjusted the missing events received from the backend, towards the destination API when making the HTTP request.

The logic was added to generate the executable application, that is, in its minified version.