Progress in the implementation of websocket requests
The models to be used for event communication throughout the life cycle of the websocket connection have been defined.

Reissue events to the server via connection tunnel.

Implementation of error handling during the connection (advanced).

Send data in Binary/Text in the message events of the websocket connection.

Implement events to create an instance of the websocket connection and obtain the active connection.